Rework blockade ID and container prefixing
Introduce a --name flag, that sets the name of the blockade instance.
This is used as the prefix of containers as well as iptables chains.

If --name is unspecified, a name is generated from the basename of the
current working directory. This approach mirrors what docker-compose
does.

Additional, introduces a container_name config setting, to allow
un-prefixed container names where desired.