Reuse a single host container for many commands
This is an optimization that improves the performance of all the host
commands (iptables, tc, etc) run by blockade. Instead of running each
command in its own container, a reusable (but ephemeral) container is
kept running and multiple commands are run in it via exec. This
container is configured to die on its own if orphaned, and will also be
periodically replaced by blockade. It is harmless for the container to
be killed (so long as no commands are executing at that moment), as
blockade will replace it on next use. Blockade kills the container upon
exit.

This change required a substantial refactor of the net module: I pulled
many of the top-level functions into helper classes. This was long
overdue, and makes it easier to inject dependencies, and write tests.